Important source code
these are the files needed for the successful implementation of the application.

app.js: this is made using node.js and and is a server code that renders only the front end features.it includes reading the available videos in the directory with their respective thumbnails, reading available books in the books directory and reading notes. essentially doing the "R" in "CRUD"

build.gradle.kts: this is the file that contains the plugins, dependancies and toolkits for the kotlin source code

Application.kt: this is the main code that is the main server of the entire application. This includes 
1:the uploading of videos, books and study notes
2:the updating of the videos, books and notes
3:the creation of the database using sqlite and also implements the CRUD operations using this server code

the app.js and the Application.kt use inter-process communication using the CORS API dependency and both servers communicate through the frontend